+++ src/heretic/am_map.c
@@ -1437,7 +1437,7 @@ static void AM_drawBackground (void)
     }
     else
     {
-    memset(I_VideoBuffer, 0, f_w*f_h*sizeof(*I_VideoBuffer));
+    memset(I_VideoBuffer, 0, (size_t)f_w*f_h*sizeof(*I_VideoBuffer));
     }
 }
